+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- add less_than_x_inbound AssetCheck
+- add final outbound checks for synopse and voxco
 
 ### Changes